refactor: building pages transformed to server components to improve loading speed
