+++ packages/fe-tw/src/app/building/[id]/cope/page.tsx
@@ -1,34 +1,28 @@
-"use client";
-
+import { fetchBuildingInfo } from "@/hooks/useBuildings/helpers";
 import { CopeView } from "@/components/Cope/CopeView";
-import { LoadingView } from "@/components/LoadingView/LoadingView";
-import { useBuilding } from "@/hooks/useBuildings";
-import React from "react";
-import { type Usable, use } from "react";
+import React, { Suspense } from "react";
 
 type Props = {
    params: Promise<{ id: string }>;
 };
 
-export default function BuildingCopePage({ params }: Props) {
-   const { id } = use<{ id: string }>(params as unknown as Usable<{ id: string }>);
-   const { data: building, isLoading } = useBuilding(id);
-
-   if (isLoading) return <LoadingView isLoading />;
+export default async function BuildingCopePage({ params }: Props) {
+   const id = await params.then((res) => res.id);
+   const building = await fetchBuildingInfo(id);
 
    if (!building) {
       return <p>Not found</p>;
    }
-
    if (!building.cope) {
       return <p>No COPE data found for {building.title}.</p>;
    }
 
    return (
-      <div className="p-4">
-         <h1 className="text-2xl font-bold mb-4">{building.title}: COPE</h1>
-
-         <CopeView cope={building.cope} />
-      </div>
+      <Suspense fallback={"Loading..."}>
+         <div className="p-4">
+            <h1 className="text-2xl font-bold mb-4">{building.title}: COPE</h1>
+            <CopeView cope={building.cope} />
+         </div>
+      </Suspense>
    );
 }
