+++ packages/fe-tw/src/app/building/[id]/liquidity/page.tsx
@@ -1,31 +1,25 @@
-"use client";
-
+import { fetchBuildingInfo } from "@/hooks/useBuildings/helpers";
 import { BuildingAddLiquidity } from "@/components/Buildings/BuildingLiquidity";
-import { LoadingView } from "@/components/LoadingView/LoadingView";
-import { useBuildings } from "@/hooks/useBuildings";
-import React, { use, type Usable } from "react";
+import React, { Suspense } from "react";
 
 type Props = {
    params: Promise<{ id: string }>;
 };
 
-export default function LiquidityPage({ params }: Props) {
-   const { id } = use<{ id: string }>(params as unknown as Usable<{ id: string }>);
-   const { buildings } = useBuildings();
-   const building = buildings.find((_building) => _building.id === id);
-
-   if (!buildings?.length || !id) {
-      return <LoadingView isLoading />;
-   }
+export default async function LiquidityPage({ params }: Props) {
+   const id = await params.then((res) => res.id);
+   const building = await fetchBuildingInfo(id);
 
    if (!building) {
       return <p>Not found</p>;
    }
 
    return (
-      <div className="p-4">
-         <h1 className="text-2xl font-bold mb-4">{building.title}: Mint and Liquidity</h1>
-         <BuildingAddLiquidity buildingAddress={building.address as `0x${string}`} />
-      </div>
+      <Suspense fallback={"Loading..."}>
+         <div className="p-4">
+            <h1 className="text-2xl font-bold mb-4">{building.title}: Mint and Liquidity</h1>
+            <BuildingAddLiquidity buildingAddress={building.address as `0x${string}`} />
+         </div>
+      </Suspense>
    );
 }
